+++ src/lib.rs
@@ -44,6 +44,5 @@ pub const INITIAL_MONOTONIC_CYCLE_COUNTER: u32 = 1024;
 
 // #[cfg(test)]
 pub mod helper;
-pub mod in_memory_data_source;
 pub mod proof_wrapper_utils;
 pub(crate) mod tests;
